+++ my-app/src/App.tsx
@@ -5,7 +5,7 @@ import Home from 'components/Main/Main';
 import Layout from 'components/Layout/Layout';
 import About from 'components/About/About';
 import NotFoundPage from 'components/NotFound/NotFoundPage';
-import FormContainer from 'components/Form/FormContainer-1';
+import FormContainer from 'components/Form/FormContainer';
 import { routes } from 'helpers/constants/routes';
 
 function App() {
