+++ my-app/src/components/Cards/Card.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { IProps } from './Cards';
-import house from '../../mockData/house.png';
+import house from '../../media/house.png';
+import fire from '../../media/fire.svg';
 
 interface CardProps {
   data: IProps;
@@ -9,7 +10,11 @@ const Card = (props: CardProps) => {
   return (
     <div className="card">
       <div>
-        <img className="card-img" src={props.data.picture ? props.data.picture : house} alt="" />
+        <img
+          className="card-img"
+          src={props.data.picture ? props.data.picture : house}
+          alt="house"
+        />
       </div>
       {props.data.address && <div>address: {props.data.address}</div>}
       {props.data.houseType && <div> type: {props.data.houseType}</div>}
@@ -22,7 +27,12 @@ const Card = (props: CardProps) => {
       {props.data.isActive !== null && <div>{props.data.isActive ? 'is on sale' : 'sold'}</div>}
       {props.data.about && <div className="features">features: {props.data.about}</div>}
       {props.data.date && <div className="">ad publication date: {props.data.date}</div>}
-      {props.data.isUrgent && <div className="">Emergency sale !!!</div>}
+      {props.data.isUrgent && (
+        <div>
+          <div className="">Emergency sale !!!</div>
+          <img className="fire" src={fire} alt="fire" />
+        </div>
+      )}
     </div>
   );
 };
