+++ my-app/src/components/Form/FormContainer.tsx
@@ -1,14 +1,22 @@
-import { IProps } from 'components/Cards/Cards';
-import React, { ChangeEvent, Component, createRef, FormEvent } from 'react';
+import React, { FormEvent, Component, createRef } from 'react';
+import { v4 as uuidv4 } from 'uuid';
+import Cards, { IData, IProps } from 'components/Cards/Cards';
 import './Form.css';
 import FormComponent from './FormComponent';
-import { v4 as uuidv4 } from 'uuid';
 
-interface IFormProps {
-  handleFormSubmit: (house: IProps) => void;
-}
-
-export interface IValidateState {
+const addressValidate = (address: string | undefined) => {
+  if (address) return address.length > 7 ? true : false;
+};
+const dateValidate = (date: string | undefined) => {
+  if (date) {
+    const year = new Date(date).getFullYear();
+    const currentYear = new Date().getFullYear();
+    if (year < 1950 || year > currentYear) return false;
+    return true;
+  }
+  return false;
+};
+interface IValidateState {
   formIsActive: boolean;
   validate: {
     id: boolean;
@@ -24,20 +32,10 @@ export interface IValidateState {
     date: boolean;
   };
 }
-const addressValidate = (address: string | undefined) => {
-  if (address) return address.length > 7 ? true : false;
-};
-const dateValidate = (date: string | undefined) => {
-  if (date) {
-    const year = new Date(date).getFullYear();
-    const currentYear = new Date().getFullYear();
-    if (year < 1950 || year > currentYear) return false;
-    return true;
-  }
-};
-
-class FormContainer extends Component<IFormProps> {
-  state: IProps & IValidateState = {
+type IState = IProps & IValidateState & IData;
+class FormPageContainer extends Component {
+  state: IState = {
+    data: [],
     id: '',
     index: 0,
     isActive: true,
@@ -65,7 +63,6 @@ class FormContainer extends Component<IFormProps> {
       date: true,
     },
   };
-
   file = createRef<HTMLInputElement>();
   date = createRef<HTMLInputElement>();
   address = createRef<HTMLInputElement>();
@@ -73,86 +70,67 @@ class FormContainer extends Component<IFormProps> {
   isActive = createRef<HTMLInputElement>();
   isUrgent = createRef<HTMLInputElement>();
 
-  handleOnChange = async (e: ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-    await this.setState((prevState: Readonly<IValidateState>) => {
-      const validate = prevState.validate;
+  handleOnChange = (e: FormEvent<HTMLInputElement | HTMLSelectElement>) => {
+    const { name } = e.target as HTMLInputElement;
+    this.setState((prevState: Readonly<IState>) => {
+      const { validate } = prevState;
       return {
-        validate: { ...validate, [e.target.name]: true },
+        formIsActive: true,
+        validate: { ...validate, [name]: true },
       };
     });
-    const arr = await Object.values(this.state.validate);
-    if (!arr.includes(false)) {
-      await this.setState({ formIsActive: true });
-    }
   };
-
-  handleOnSubmit = async (e: FormEvent) => {
+  handleOnSubmit = (e: FormEvent) => {
     e.preventDefault();
-    const address = this.address.current?.value;
-    if (!addressValidate(address)) {
-      await this.setState((prevState: Readonly<IValidateState>) => {
-        const validate = prevState.validate;
+    this.setState((prevState: Readonly<IState>) => {
+      const address = this.address.current?.value;
+      const date = this.date.current?.value;
+      const houseType = this.houseType.current?.value;
+      const isActive = !this.isActive.current?.checked;
+      const isUrgent = this.isUrgent.current?.checked;
+      const current = this.file.current as HTMLInputElement;
+      const { files } = current;
+      const picture = files != null && files[0] ? URL.createObjectURL(files[0]) : '';
+      const validate = prevState.validate;
+      const data = prevState.data;
+      if (!addressValidate(address) || !dateValidate(date) || !addressValidate(houseType)) {
         return {
+          validate: {
+            ...validate,
+            address: addressValidate(address),
+            date: dateValidate(date),
+            houseType: addressValidate(houseType),
+          },
           formIsActive: false,
-          validate: { ...validate, address: false },
         };
-      });
-    }
-    const date = this.date.current?.value;
-    if (!dateValidate(date)) {
-      await this.setState((prevState: Readonly<IValidateState>) => {
-        const validate = prevState.validate;
-        return {
-          formIsActive: false,
-          validate: { ...validate, date: false },
-        };
-      });
-    }
-    const houseType = this.houseType.current?.value;
-    if (!addressValidate(houseType)) {
-      await this.setState((prevState: Readonly<IValidateState>) => {
-        const validate = prevState.validate;
-        return {
-          formIsActive: false,
-          validate: { ...validate, houseType: false },
-        };
-      });
-    }
-
-    const arr = await Object.values(this.state.validate);
-    if (arr.includes(false)) return;
-    const isActive = !this.isActive.current?.checked;
-    const isUrgent = this.isUrgent.current?.checked;
-    await this.setState({
-      id: uuidv4(),
-      houseType,
-      isUrgent,
-      isActive,
-      address,
-      date,
-    });
-    const current = this.file.current as HTMLInputElement;
-    const { files } = current;
-    await this.setState(() => {
-      if (files != null && files[0]) {
-        return { picture: URL.createObjectURL(files[0]) };
-      } else {
-        return { picture: '' };
       }
+      return {
+        data: [
+          ...data,
+          {
+            id: uuidv4(),
+            houseType,
+            isUrgent,
+            isActive,
+            address,
+            date,
+            picture,
+          },
+        ],
+      };
     });
-    await this.props.handleFormSubmit(this.state);
   };
-
   render(): React.ReactNode {
     return (
       <FormComponent
+        data={this.state.data}
         handleOnSubmit={this.handleOnSubmit}
         handleOnChange={this.handleOnChange}
-        address={this.address}
-        addressError={!this.state.validate.address}
-        dateError={!this.state.validate.date}
-        houseTypeError={!this.state.validate.houseType}
+        addressValidate={this.state.validate.address}
+        dateValidate={this.state.validate.date}
+        houseTypeValidate={this.state.validate.houseType}
         formIsActive={this.state.formIsActive}
+        address={this.address}
         houseType={this.houseType}
         isActive={this.isActive}
         isUrgent={this.isUrgent}
@@ -163,4 +141,4 @@ class FormContainer extends Component<IFormProps> {
   }
 }
 
-export default FormContainer;
+export default FormPageContainer;
